added Redux for <CoinSearcher />
