+++ src/store/models/homePage.js
@@ -10,10 +10,14 @@ import * as tradeHistoryActionCreators from '../actions/tradeHistory';
 import * as orderHistoryActionCreators from '../actions/orderHistory';
 import * as depthChartActionCreators from '../actions/depthChart';
 
+// import * as orderHistoryActionCreators from '../actions/orderHistory';
+import * as coinSearcherActionCreators from '../actions/coinSearcher';
+
 import * as orderList from '../../jsons/ordersList.json';
 import * as tradeHistory from '../../jsons/tradeHistory.json';
 import * as orderHistory from '../../jsons/orderHistory.json';
 import * as bidAsk from '../../jsons/bidAsk.json';
+import * as coinsList from '../../jsons/coinsList.json';
 
 export default function getHomePageModel(state: State) {
   return HomePageModel(state.homePage);
@@ -44,6 +48,13 @@ const depthChartData = {
   data: bidAsk.list,
   loading: false,
   title: 'ETJ/BTC',
+}
+const coinSearcherData = {
+  coinsList: coinsList.list,
+  loading: false,
+  small: false,
+  decimals: 7,
+  authenticated: true,
 };
 
 export const loadData = ({ tokenId }: LoadDataParams): ThunkAction => {
@@ -55,5 +66,6 @@ export const loadData = ({ tokenId }: LoadDataParams): ThunkAction => {
     dispatch(tradeHistoryActionCreators.saveData(tradeHistoryData));
     dispatch(orderHistoryActionCreators.saveData(orderHistoryData));
     dispatch(depthChartActionCreators.saveData(depthChartData));
+    dispatch(coinSearcherActionCreators.saveData(coinSearcherData));
   };
 };
