+++ src/types/index.js
@@ -10,6 +10,7 @@ import type { HomeState } from './homePage';
 import type { TradeHistoryState } from './tradeHistory';
 import type { DepthChartState } from './depthChart';
 import type { OrderHistoryState } from './orderHistory';
+import type { CoinSearcherProps } from './coinSearcher';
 import type { TokenState, TokenAction, TokenEvent } from './tokens';
 import type { AccountState, AccountAction, AccountEvent } from './account';
 import type { DepositFormState, DepositFormAction, DepositFormEvent } from './depositForm';
@@ -54,6 +55,7 @@ export type State = {
   tradeHistory: TradeHistoryState,
   orderHistory: OrderHistoryState,
   depthChart: DepthChartState,
+  coinSearcher: CoinSearcherProps,
   etherTx: EtherTxState,
   tokens: TokenState,
   account: AccountState,
